MB-31101: Basic collection support for getRandomKey
Look for a random key from a collection, not the bucket. As there
is no in-memory index for a collection, we currently iterate through
memory.

This commit only covers the ep-engine side of the change, there is
no way yet for the client to encode a collection into their request,
so for now lock into the default collection.

The lack of an index in-memory for each collection leads the code
worst case search all keys in memory.

Some quick timings to see what this is like where we have 1 key in the
default collection (which this commit is locked into searching) and
n keys in a second collection.

keys in           | average 'us' to find
second collection | default collection key

-------------------------------------------
       1          |   130
       1,000,00   |   224
       11,000,000 |  1931
       13,500,000 |  2947

Change-Id: Ifd45c9504059c5f763067f7472d5ce795b414a0b
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/128125
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>