MB-35889: Don't invalidate index entry for Disk checkpoint for expel
When we are expelling we attempt to invalidate the index entry as the
pointer will no longer be valid. This is not valid for Disk checkpoints
anymore as we do not put keys into the indexes of Disk checkpoints.

Do not attempt to invalidate index entries for Disk checkpoints when
expelling, or attempt to make any required assertions.

Change-Id: I9fe761982cda9c5aebfd1f28358317154120de01
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/128352
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>