MB-39416: Cleanup files in test_whitespace_db properly
Currently this test fails sometimes as files are not cleaned up before
or after the test. Cleanup the files properly before (in case we crash)
and after the test runs.

Remove the dbname parameter from the specific test config as this value
is never used and is misleading. Due to the way that ep_testsuite works
it is overriden by the dbname from the command line.

Change-Id: I4db8518e6807d2e971ce776ac782c338f60e6420
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/128616
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>