MB-39441: Don't insert multiple error entries
A subdoc multi mutation is expected to return the error
information for the _first_ error it hit (as execution
is terminated at that state).

As part of adding support for XAttrs we ended up adding
one error entry for each phase which would get the protocol
out of sync (as the header only allocates room for one).

Change-Id: Ic7cb9fb8c83aeb9d70566360543c2d954d87024c
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/128826
Tested-by: Build Bot <build@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Daniel Owen <owend@couchbase.com>