MB-36371: Allow magma to purge tombstones in implicit compactions
Currently magma can only purge tombstone when an explicit compaction is
run. Magma needs to be able to purge tombstone during implicit
compactions as these will be the main types of compactions run for a
magma bucket. Allow magma to purge tombstones in implicit compactions
by setting the metadata purge ts when we create a compaction_ctx for
magma.

MB-39311: Do not consider purge seqno when compacting magma tombstones
Normally tombstones are only purge when their seqno is less than the
purge seqno to try to prevent Views from rolling back. This value is
sent in every CompactDB payload by ns_server and there is no other
mechanism to track this for magma buckets. Discussion is still ongoing
about the long term solution, but in the short term we should allow
magma to purge tombstones without regard to the purge seqno for
functional completeness (milestone 1).

These two changes are being made in the same patch as the former
requires that we ignore the purge seqno in some regards (implicit
compactions) and so it makes sense to simplify the logic and ignore
the purge seqno in all regards (the latter change).

Change-Id: Ifddd9a3246dc49f49e4d1a6750327df375b6b502
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/128861
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>