MB-39310: Change magma fragmentation ratio to percentage
Magma fragmentation ratio was added recently as a ratio as this is
the form that magma expects the value to be passed in. ns_server does
not play nicely with float KV configuration (we are returning 0 instead
of 0.0 and other rounding/truncation errors may be an issue). Make
magma fragmentation ratio an integer percentage instead to "fix" this.
Furthermore, magma sets a minimum of 0.1 (10%) for their fragmentation
ratio so impose a minimum value of 10.

Change-Id: Ib5d086bc98e4e056b06ca928a462ff92263d93d0
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/129034
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Sarath Lakshman <sarath@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>