MB-39549: Move begin to KVStore and clear collections commit metadata
Currently the collections commit meta is only cleared on success
and that leads to an exception if commit fails and retries (when
the commit contains collection state changes).

Reviewing the implementations of begin shows they are nearly all
identical, couch-kvstore has more error checking but they all
result in the same output - transation bool = true and move of
the input parameter.

All KVStores need to handle the collections metadata too, so on
begin we can clear the state ready for whatever the commit batch
may contain.

Change-Id: I78f06aed32b37223c22480bec12ae35527b4ae19
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/129159
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>
Reviewed-by: Ben Huddleston <ben.huddleston@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>