MB-37680: Schedule backfill of snap_markers ASAP
When the BackfillManager sequential schedule order was added it
resulted in the second (and subsequent) scheduled vbucket backfills
not running at all until the first was completed. While this might
been desirable, it has the effect of delaying when the create phase of
the backfill occurs. This means that the actual opening of the
couchstore file (or initialization of range read for ephemeral) is
delayed until the earlier vBuckets have completed.

As a consequence, the consistency of different disk-only backfills
(the time difference between the high seqno in the snapshot) is
increased, potentially very significantly.

For example, if a DCP connection requests backfill on 4 vBuckets, and
each vBucket takes 1 minute to fully backfill (not unreasonable for a
large dataset), then when using Sequential schedule order the opening
of the second and subsequent vBuckets will be delayed by:

    T=0    Open vb0, create Snapshot Marker(vb0), send Mutation(vb0...)
    T=60s  Open vb1, create Snapshot Marker(vb1), send Mutation(vb1...)
    T=120s Open vb2, create Snapshot Marker(vb2), send Mutation(vb2...)
    T=180s Open vb3, create Snapshot Marker(vb3), send Mutation(vb2...)

Here there would be 180s skew between vb0's snapshot and vb3's.

To address this, the BackfillManager 'sequential' schedule order is
slightly relaxed - the first step of new Backfills is always
round-robin, only once the Backfill has been run once does the
scheduleOrder parameter take effect.

Change-Id: If85dc61552f3ab52eccbd9373a37c6eaf1dafbb6
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/129292
Reviewed-by: Ben Huddleston <ben.huddleston@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>