MB-39570: Advertise all configured SASL mechanisms
Always advertise all of the configured authentication
mechanisms even if they're not possible for all users. This
allows local users to use SCRAM-SHA (which won't work for
a user defined in LDAP).

The server adds information to the error context if the
server is configured with an external auth provided to
to give the user a "hint" of a possible reason why
the authentication failed (if the user did provide
the correct credentials)

Change-Id: I82ad07971d66193d47e0dd84eee8b9031511ae02
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/129125
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>