MB-37009: DcpProducer::handleResponce, don't disconnect on NotMyVbucket
Before this patch we will disconnect the connection between active to
replica if we see the replica (DcpConsumer) return a NotMyVbucket. This
is problematic as we don't want to stop replication for all vbuckets
just because one vbucket has been removed on the consumer side, and its
likely that the stream is about to go away soon anyway. Thus, when a
consumer receives an unexpected NotMyVbucket we should log it as an
error but we should not disconnect the connection.

Change-Id: I57ec4d2ca4b9227d5ce9dd0cde713d8e66b6a186
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/129383
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>