MB-39280: Flusher marks-clean item in the HashTable by Seqno
We used to check the CAS in place of Seqno in the PersistenceCallback,
which is wrong as we have legal scenarios where multiple mutations with
different Seqno and same CAS (for the same key) may be queued in the
system.

Eg, an XDCR setup where multiple source clusters replicate to the same
destination cluster.
In that case, we may end up with marking clean (by CAS) a StoredValue
that has never been persisted. That makes the SV eligible for eviction.
Which in turn may lead to exposing a stale item from disk.

See the unit test linked to this fix for a detailed reproduction of the
described scenario.

Change-Id: I8873ea3b4078284c02c3f17dec7ae20df0ee122b
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/129399
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>