MB-39375: Add Setting for storage thread
Add a new memcached setting "num_storage_threads" which is a global
configuration value that we can pass through to the engine to set the
appropriate number of backend storage threads. This configuration
parameter will have a default value that can be supplied by passing a
configuration parameter of 0.

In a following patch we will add a "magma_flusher_thread_percentage"
configuration parameter than is required to calculate the appropriate
number of magma flusher and compactor threads.

Change-Id: Iae43f14ef914c8b8cb02d996dcdede650614f2a9
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/129840
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>