MB-39456: Refactor GetKeys code in ep_engine.cc to getkeys.(h/cc)
Refactor code for GetKeys in ep_engine.cc by moving it to getkeys.(h/cc)
so that we can separate out the get keys task and callback out of
ep_engine.cc. Which is an very large file, of which very little code
relates to the get keys functionality of ep engine.

Change-Id: I43ec31cf0007c3762df289ea0940fafc5cb4d35c
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/129590
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>