MB-39650: DCP input does not apply to collection RBAC
Incoming DCP changes to a bucket (mutation/deletion etc...) should only
be privilege checked for bucket level Privilege::DcpConsumer, we can
not express hierarchical (scope/collection) privileges for incoming
replication (the bucket will fail to be a replica). So for the
is_collection_opcode function which is used to determine if the
mcbp front end should extract scope/collection identifiers for input
to mcbp_privileges we return false for DcpMutation/DcpDeletion etc...
This means they can only ever be checked against the bucket for
DcpConsumer.

The major side effect of this change is that we bypass the possibility
for mcbp to fail DCP input for NotMyVbucket (which is an expected error
for rebalance) and all DCP input will end-up in the engine's consumer
code - meaning that engine side functionality - DCP flow control and
DCP noop/keep-alive doesn't cause pauses or disconnects.

A note on testing - we don't have unit testing infrastructure that
would demonstrate the issue/fix, full stack DCP is required and
the ability to reliably interleave vb state changes with traffic.

Change-Id: Ibc48747e8ebcb40f5e5f39c7d231ad2c0a9a106f
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/129556
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>