MB-39620: Document new ifconfig commands [1/3]
Interface definitions should not be part of memcached.json
as we don't have a way to report back to ns_server if an error
occurs while trying to bind to one (or more) of the new ports
in the new configuration. In theory we could end up in a situation
where no one knows how to connect to memcached as it failed to
bind to any of the ports in the new configuration and continues
to run with the previous configuration (but no one knows that).

To solve the problem memcached will create "bootstrap interfaces"
which use ephemeral ports and is only bound to the loopback interface.
For the controlling process to know how to connect to these ports they
are stored in the "port number file" provided in the configuration.

After startup the controlling process must use the ifconfig command
to create/delete interfaces (and these interface definitions are
NOT preserved across process restarts).

The patch series is split into 3 parts:

1 - Document the new feature (this patch)
2 - Add support for the new feature
3 - Remove the support for interface and prometheus in
    memcached.json

Change-Id: I04fb09b856ac7469815201bcb413a9d2190af0f6
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/129583
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>