MB-34267: Add retry logic to reconfigure
MaxConnectionTest::SetUp tries to reconfigure the server
to tune the connection settings to avoid having to create
a ton of connections before it hits the max limits.

reconfigure() is called, but it might race with the
internal counters on the server on how many connections
the server currently have (as different threads closing
connections may not have been scheduled to run). To
work around this the patch adds retry logic in
TestappTest::reconfigure so any exception thrown during
recinfigure cause a retry. To avoid running "forever"
a hard timeout is set to 5 seconds.

Change-Id: Iaaf2686b813d56465467732e85f9d670fbd1e1c6
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/129600
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Daniel Owen <owend@couchbase.com>