Refactor: Rename dispatcher to network interface manager
Get rid of the "dispatcher thread" object and move the logic
into a new class called NetworkInterfaceManager which is a
better name for what it is responsible of doing. Currently
the event_handler is still stuck in memcached.cc as most
of the method it is calling is still located there.

Change-Id: Ie56c6a4d24712d2ab13588c1f46c85f778b8c072
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/129701
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>