MB-34217: Add prepare param to EphemeralVB::isKeyLogicallyDeleted
In EphemeralVBucket::isKeyLogicallyDelete() we call dropKey() to drop
a key that has been deleted due to a collection drop. We need to be
able to drop prepares from the DM's here in a future patch and so we
need to know whether or not we are working with a prepare.

Change-Id: I44cc6220cb9f0584c0b692d6895586d5be939648
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/129838
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>