Fix get_scope_id to make sure it returns default scopeID
Currently if EventuallyPersistentEngine::get_scope_id() will return
cb::engine_errc::unknown_collection, when passed a DocKey that is
in the default scope and is not given a Vbid. This is due to an error
in the use of the std::optional that wraps the ScopeID returned by
KVBucket::getScopeID(). Where we would check if we should return the
scopeID by observing the scopeID rather than checking if the
std::optional had a value. This causes problems in the default scope
case as its scopeID is 0 and thus, when used in a boolean condition
results in false.

To fix this ensure we use std::optional's has_value() method and also
add unit tests to verify its behavior.

Change-Id: I8a540cdb23887a416cbfe050f0522be044708c19
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/129923
Reviewed-by: Jim Walker <jim@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>