MB-37009: DcpProducer::handleResponse, don't disconnect on KeyEexists
DcpProducer shouldn't disconnect on status code
cb::mcbp::Status::KeyEexists. As this is returned by
DcpConsumer::lookupStreamAndDispatchMessage(), when the stream found
for the op's vbucket has an opaque that does not match the op's
opaque. This can occur during rebalance due to stream takeover and
thus we don't want to tear down all the streams because it's likely a
vbucket is being moved to another node or has changed state.

Change-Id: I6ed38cd9b50b40b84b0a354b2a26f952bb380d80
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/130046
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: James Harrison <james.harrison@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>