MB-36335: Save local documents with cb::couchstore::saveLocalDocuments
Replace the use of many couchstore_save_local_document calls with
one call to cb::couchstore::saveLocalDocuments. This is done by
creating a second queue for the local index key/value pairs, having
the existing functions which write to the local index write to the
queue and finally draining the queue before we commit.

Change-Id: I106caa557ac2cfda789b2734c59ed795559f9c1f
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/130126
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>