MB-36335: Save local documents with cb::couchstore::saveLocalDocuments
Replace the use of many couchstore_save_local_document calls with
one call to cb::couchstore::saveLocalDocuments.

There are a number of paths (executed in a few concurrent contexts)
that update local documents.

For the flusher begin()/set|del/commit() path a queue exists attached
to CouchKVStore and CouchKVStore::commit writes out all of the updates
added to that queue before couchstore_commit

For compactInternal a queue is created on the stack of the
function, only two updates can actually occur and they use
the locally scoped queue so that they cannot interfere with the
flusher path.

Finally writeVBucketState can also update just the vbstate document,
this doesn't use new saveLocalDocuments API, but just the existing
single update method.

Change-Id: Ic92e87c300be68c964ee184175985169dedd90eb
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/130336
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>