MB-39717: Make magma collections compaction compact prepare namespace
Currently the magma collections drop compactions trigger a range
based compaction over each dropped collection. This does not consider
the prepare namespace which also needs to be visited when dropping
collections. This needs to be done separately as the prepare namespace
prefix precedes the collection prefix in the on disk key.

Change-Id: Ib59c964ac88d65080d85c7f6e9a229b6ea915a78
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/130212
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>