Remove Tracer::end which search for span id
The caller should keep track of the Span ID it is using and
update that when calling end. The version which took
a trace code would not work if the same trace code was
used multiple times (ex: if we got kicked off the CPU
and ended a span and wanted to start a new span with
the same code at a later time).

Change-Id: I9832a4d4c4baef882356f4aae265b966d23c5196
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/130612
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>