MB-16181: Improve leb128 input checking
Before this change there were inputs which would of decoded without
error, but are invalid. E.g. the following u32int leb128

    0x81.82.83.84.85.86.87.00  (8 bytes)

Passed into the uint32_t decoder (which is what KV-engine uses)
would of produced a result, however it is invalid input. The
decoder loop would of been shifting by greater than 32 for some
of the bytes (and ASAN complains about that, which lead to this
commit).

This input should of been detected as invalid. Each leb128 type
has a maximum length, for u32int 5 bytes is the maximum encoded
length.

This patch uses the existing getMaxSize() constant to ensure that
the decode function only consumes up to the maximum number
of bytes The buffer itself can still be larger, but the decoder
stops for a stop-byte or the max. if no-stop byte is found in the
leb128 bytes, the input is invalid and an exception or detectable
error return is produced.

Change-Id: I83622d26067cbe9ffdfbbd86bd39d1b5077f217d
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/130755
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: James Harrison <james.harrison@couchbase.com>