MB-40052 [2/3]: Introduce BackfillTrackingIface
Further decouple BackfillManager from DcpConnMap, by introducing a
BackfillTrackingIface. This interface allows the implementer to track
the active Backfills, and is used by BackfillManager to know when to
place Backfills onto the pending list.

By adding this interface it simplifies BackfillManager unit tests - no
need to mock the entire DcpConnMap if behaviour of it needs to be changed.

Change-Id: I9999a2a2b9ea8e9a720bcf8d6c4d3b2c82fa15aa
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/130975
Tested-by: Build Bot <build@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: James Harrison <james.harrison@couchbase.com>