MB-39297: Correct magma VB doc count underflow
Currently the VB doc count can underflow for magma vBuckets when we
drop collections as we incorrectly track the number of items purged
during a collection erase (compaction). This happens because magma
can call the compactionCallback multiple times for a given key as
it may store old values for each keys until compaction cleans them
up.

Correct the doc count underflow by using the collection item count
stats instead of the diff stat calculated during the compaction. This
required postponing the deleting of the stats local doc until the
compaction runs for magma instead of deleting the stats local doc when
we flush the deletion of the collection.

Change-Id: I4cafecef4aecc528a399b9aec9dc95cd02bd5c36
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/130973
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>