MB-40052 [3/3]: Account for Backfills in initializingQ on destruction
When Backfills are started (moved into the initializingQ) the
BackfillManager notifies the DcpConnMap so it can track how many
Backfills across the whole bucket are in progress, and if necessary
cap the number which can concurrently run.

When these backfills complete (either successfully or are cancelled),
then the BackfillManger needs to also notify DcpConnMap to decrement
the number in-progress.

However, when the additional initializingQ was added to
BackfillManager for MB-37680, Backfills in that queue were *not*
decremented from the number in progress.

The effect of this was such Backfills effectively "leaked", meaning
that with sufficent number DCP backfilling for the entire bucket could
hang.

Fix by adding the missing accounting.

Add tests to check that active or snoozing backfills which are still
in existence when a BackfillManager is destructed, are correctly
accounted in the BackfillTracker.

Change-Id: I00215072e9558e7f5cdcd1672f800027d90124ac
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/130995
Tested-by: Build Bot <build@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: James Harrison <james.harrison@couchbase.com>