MB-39594: Reduce work required for hash-table key compare
Key comparison in the hash-table uses SerialisedDocKey::hasKey (and
the operator==). The compare worked by decoding the collection from
both operands, then comparing 1) collection, 2) the logical key size
and finally the key data.

KV-engine now only accepts canonical collection encodings which means
comparing DocKey (and sub-classes) can be done more like a memcmp.
1) check size 2) check data

We just have to check that the DocKey (rhs) operand does or does not
encode a collection-ID and adjust the comparison to account for that
case.

DurabilityPrepare: The code being replaced would always compare the keys
as they are, e.g. A prepared "key" in collection X would not match the
"key" in collection X, this updated logic maintains that.

Change-Id: Ib912851992ccae9170c67e46272750ec3403f726
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/131178
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>