MB-39594: Test for 'System' collection with 1 byte load/compare
Profiling a collection enabled workload showed functions taking
a higher percentage compared to a non-collection workload. This
MB aims to reduce the following:

* Item::Item called from
 * EventuallyPersistentEngine::itemAllocate
 * StoredValue::toItemBase

Item::Item is decoding the collection of the key and then checking
if it is CollectionID::System. With LEB128 encoding there is no
need to do the full decode when testing for values <= 127, byte 0
of the key can be inspected directly.

Change-Id: I37cee0468fdf84496d806e566e506a3945eb3fa5
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/130412
Reviewed-by: James Harrison <james.harrison@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>