MB-40095: Ensure DcpSeqnoAdvanced is not sent when SyncReplication is on
Ensure we never send a DcpSeqnoAdvanced is never sent when
SyncReplication is enabled on the DCP connection. Also refactor how we
decide to enqueue a SeqnoAdvanced, to make it clear to the reader.

Also add two tests to ensure we do/do not send a DcpSeqnoAdvanced when
SyncReplication is disabled and enabled.

Change-Id: Ib4c0479f2a1af207c76dfed3ee01ca5f790a8b6f
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/131270
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>