MB-39532: Tolerate a collection being re-created
Remove an exception from the collection state flusher so that we
allow a collection to be in both drop and open lists.

This has the effect of fixing crashes seen in the MB but is also
how flush of a collection is intended to work.

Collections have an associated seqno (the seqno at which they were
created). The isLogicallyDeleted function which decides if a key
is dropped (because the collection was dropped) has always used the
collection-ID and seqno for evaluating the result. In the flush case
a collection's start seqno would effectively move to vbucket-high
meaning all existing document's of the collection are dropped, but
new documents can be added. With this always being the design, the
exception was incorrect and can be removed.

Change-Id: I5a67af818c0dbba7a3d3f29e56cc405df7173d4e
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/131375
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Richard de Mellow <richard.demellow@couchbase.com>