MB-39598: Reduce iteration count of ep_testsuite_basic 'multi set'
This test intermittently times out on the Magma storage backend. The
intrinsic performance of Magma vs Couchstore is being tracked via
MB-40105, but to avoid blocking CV, avoid this issue by droping the
iteration count of this test by 5x - note this is primarily a
functional test and hence 20000 items is perfectly sufficient.

Change-Id: Ie14ceda3049815cdc4616284060b7d0e208bd57a
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/131422
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>