MB-40145: Remove manifest-uid filtering of DCP system events
The DCP filter code is using the client's manifest-uid (which would be
set on stream resumption) to decide if an event should be sent. This may
well work fine if everyone follows the protocol (using the last
manifest-uid observed on the stream), however it  1) seems unnecessary,
the existing start-seqno is enough to resume a client, all events after
that seqno should be sent (profiled they match
scope/collection/pass-through).

2) There seems to be no reason for it (no test case and original commit
gave no justification for the logic behind the hiding)

We should only be using the manifest-UID to prevent a situation where a
fail-over could lead to a client from the 'future' DCP  connecting,
which remains in-place.

Change-Id: I82989824ce30f64de2c879dcda7a82eaacbdebbf
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/131434
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Ben Huddleston <ben.huddleston@couchbase.com>